backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-multipart
rembg
opencv-python-headless
pillow
ultralytics
numpy
requests
